outsight/requirements.txt
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
conda==4.11.0
conda-package-handling==1.7.3
cryptography==3.4.7
d6tflow==0.2.5
idna==2.10
imageio==2.14.1
imutils==0.5.4
# menuinst==1.4.16
# mkl-fft==1.3.1
# mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.6.3
numexpr==2.8.1
numpy==1.21.2
# opencv-contrib-python==4.5.5.62 # is only pip
packaging==20.4
pandas==1.3.5
Pillow==9.0.0
pip==21.1.3
pycosat==0.6.3
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
pytesseract==0.3.8
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.2.0
# pywin32==228
pyyaml==6.0
requests==2.25.1
# ruamel-yaml-conda==0.15.100
scikit-image==0.19.1
scipy==1.7.3
setuptools==58.0.4
six==1.16.0
tifffile==2021.11.2
tqdm==4.61.2
urllib3==1.26.6
wheel==0.36.2
# win-inet-pton==1.1.0
# wincertstore==0.2